Fix for errors reported during testingand memory leaks.
This commit includes
 - Fixing memory leaks
 - Fixing coverity defects
 - Fixing errors reported in testing related to lifetime parameter.

+++ test/lwm2mtestapplication/lwm2mtest.cpp
@@ -29,7 +29,7 @@ M2MLWClient::~M2MLWClient()
 
 bool M2MLWClient::create_interface(const char *endpoint,
                       const char *resource_type,
-                      const uint16_t lifetime,
+                      const int32_t lifetime,
                       const uint16_t listen_port,
                       const char *domain,
                       const uint8_t binding_mode,
